pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "rawnind"
version = "0.2.0"
description = "RAW imaging and pipelines for neural datasets"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Rawnind Maintainers" }]
requires-python = ">=3.12"
dependencies = [
    "black>=25.9.0",
    "colour==0.1.5",
    "colour-science==0.4.6",
    "configargparse==1.7.1",
    "docker-py>=1.10.6",
    "exif==1.6.1",
    "httpx>=0.28.1",
    "matplotlib>=3.10.3",
    "notebook>=7.1.3",
    "opencv-python==4.12.0.88",
    "openexr==3.3.5",
    "openimageio==3.0.6.1",
    "piexif==1.1.3",
    "piqa==1.3.2",
    "psutil>=7.0.0",
    "ptflops==0.7.5",
    "pylint==3.3.8",
    "pypng==0.20220715.0",
    "pytest==8.4.1",
    "pytest-trio>=0.8.0",
    "pytorch-msssim==1.0.0",
    "pyyaml==6.0.2",
    "rawpy==0.25.1",
    "requests==2.28.1",
    "rich>=14.1.0",
    "scipy-stubs==1.16.1.1",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "tqdm==4.66.5",
    "trio>=0.31.0",
]


[dependency-groups]
dev = [
    "memory-profiler>=0.61.0",
    "py-spy>=0.4.1",
]
test = [
    "pydantic>=2.12.0",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pytest-trio>=0.8.0",
    "scikit-image>=0.25.2",
    "starlette>=0.48.0",
    "tifffile>=2025.10.4",
]

[tool.pytest.ini_options]
# Only look for tests in the canonical tests/ folder (ignore tests.old)
testpaths = ["tests"]
# Ensure the src layout is importable when running tests without installation
pythonpath = ["src"]
# Common markers used in this repository
markers = [
    "trio: trio-based asynchronous tests",
    "dataset: dataset pipeline tests (DataIngestor, Downloader, FileScanner, etc.)",
]


# Setuptools configuration for src/ layout packaging
[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
include = ["rawnind*"]


[project.optional-dependencies]
dev = [
    "memory-profiler>=0.61.0",
    "py-spy>=0.4.1",
]

test = [
    "pydantic>=2.12.0",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pytest-trio>=0.8.0",
    "scikit-image>=0.25.2",
    "starlette>=0.48.0",
    "tifffile>=2025.10.4",
]
